Split state handler into groups
