+++ packages/extension-base/src/background/handlers/State.ts
@@ -3,15 +3,17 @@
 // of the Apache-2.0 license. See the LICENSE file for details.
 
 import { knownMetadata, addMetadata } from '@polkadot/extension-chains';
-import { ProviderMeta, MetadataDef } from '@polkadot/extension-inject/types';
-import { JsonRpcResponse, ProviderInterface, ProviderInterfaceCallback } from '@polkadot/rpc-provider/types';
-import { AccountJson, AuthorizeRequest, MetadataRequest, RequestAuthorizeTab, RequestRpcSend, RequestRpcSubscribe, RequestSign, ResponseRpcListProviders, ResponseSigning, SigningRequest, RequestRpcUnsubscribe } from '../types';
+import { MetadataDef } from '@polkadot/extension-inject/types';
+import { AccountJson, AuthorizeRequest, MetadataRequest, RequestAuthorizeTab, RequestSign, ResponseSigning, SigningRequest } from '../types';
+import { Providers } from './types';
 
 import { BehaviorSubject } from 'rxjs';
 import { assert } from '@polkadot/util';
 
 import chrome from '../../chrome';
 import { MetadataStore } from '../stores';
+import StateRpc from './StateRpc';
+import { getId, stripUrl } from './util';
 
 interface AuthRequest {
   id: string;
@@ -38,15 +40,6 @@ interface MetaRequest {
   url: string;
 }
 
-// List of providers passed into constructor. This is the list of providers
-// exposed by the extension.
-type Providers = Record<string, {
-  meta: ProviderMeta;
-  // The provider is not running at init, calling this will instantiate the
-  // provider.
-  start: () => ProviderInterface;
-}>
-
 interface SignRequest {
   account: AccountJson;
   id: string;
@@ -55,8 +48,6 @@ interface SignRequest {
   reject: (error: Error) => void;
   url: string;
 }
-
-let idCounter = 0;
 
 const WINDOW_OPTS = {
   // This is not allowed on FF, only on Chrome - disable completely
@@ -69,24 +60,14 @@ const WINDOW_OPTS = {
   width: 480
 };
 
-function getId (): string {
-  return `${Date.now()}.${++idCounter}`;
-}
-
-export default class State {
+export default class State extends StateRpc {
   readonly #authUrls: AuthUrls = {};
 
   readonly #authRequests: Record<string, AuthRequest> = {};
 
   readonly #metaStore = new MetadataStore();
 
-  // Map of providers currently injected in tabs
-  readonly #injectedProviders: Map<chrome.runtime.Port, ProviderInterface> = new Map();
-
   readonly #metaRequests: Record<string, MetaRequest> = {};
-
-  // Map of all providers exposed by the extension, they are retrievable by key
-  readonly #providers: Providers;
 
   readonly #signRequests: Record<string, SignRequest> = {};
 
@@ -98,8 +79,8 @@ export default class State {
 
   public readonly signSubject: BehaviorSubject<SigningRequest[]> = new BehaviorSubject<SigningRequest[]>([]);
 
-  constructor (providers: Providers = {}) {
-    this.#providers = providers;
+  constructor (providers?: Providers) {
+    super(providers);
 
     this.#metaStore.all((key: string, def: MetadataDef): void => {
       if (key.startsWith('metadata:')) {
@@ -162,7 +143,7 @@ export default class State {
       const isAllowed = result === true;
       const { idStr, request: { origin }, url } = this.#authRequests[id];
 
-      this.#authUrls[this.stripUrl(url)] = {
+      this.#authUrls[stripUrl(url)] = {
         count: 0,
         id: idStr,
         isAllowed,
@@ -193,14 +174,6 @@ export default class State {
 
       fn(result);
     };
-  }
-
-  private stripUrl (url: string): string {
-    assert(url && (url.startsWith('http:') || url.startsWith('https:')), `Invalid url ${url}, expected to start with http: or https:`);
-
-    const parts = url.split('/');
-
-    return parts[2];
   }
 
   private updateIcon (shouldClose?: boolean): void {
@@ -238,7 +211,7 @@ export default class State {
   }
 
   public async authorizeUrl (url: string, request: RequestAuthorizeTab): Promise<boolean> {
-    const idStr = this.stripUrl(url);
+    const idStr = stripUrl(url);
 
     if (this.#authUrls[idStr]) {
       assert(this.#authUrls[idStr].isAllowed, `The source ${url} is not allowed to interact with this extension`);
@@ -264,7 +237,7 @@ export default class State {
   }
 
   public ensureUrlAuthorized (url: string): boolean {
-    const entry = this.#authUrls[this.stripUrl(url)];
+    const entry = this.#authUrls[stripUrl(url)];
 
     assert(entry, `The source ${url} has not been enabled yet`);
     assert(entry.isAllowed, `The source ${url} is not allowed to interact with this extension`);
@@ -299,74 +272,6 @@ export default class State {
 
   public getSignRequest (id: string): SignRequest {
     return this.#signRequests[id];
-  }
-
-  // List all providers the extension is exposing
-  public rpcListProviders (): Promise<ResponseRpcListProviders> {
-    return Promise.resolve(Object.keys(this.#providers).reduce((acc, key) => {
-      acc[key] = this.#providers[key].meta;
-
-      return acc;
-    }, {} as ResponseRpcListProviders));
-  }
-
-  public rpcSend (request: RequestRpcSend, port: chrome.runtime.Port): Promise<JsonRpcResponse> {
-    const provider = this.#injectedProviders.get(port);
-
-    assert(provider, 'Cannot call pub(rpc.subscribe) before provider is set');
-
-    return provider.send(request.method, request.params);
-  }
-
-  // Start a provider, return its meta
-  public rpcStartProvider (key: string, port: chrome.runtime.Port): Promise<ProviderMeta> {
-    assert(Object.keys(this.#providers).includes(key), `Provider ${key} is not exposed by extension`);
-
-    if (this.#injectedProviders.get(port)) {
-      return Promise.resolve(this.#providers[key].meta);
-    }
-
-    // Instantiate the provider
-    this.#injectedProviders.set(port, this.#providers[key].start());
-
-    // Close provider connection when page is closed
-    port.onDisconnect.addListener((): void => {
-      const provider = this.#injectedProviders.get(port);
-
-      if (provider) {
-        provider.disconnect();
-      }
-
-      this.#injectedProviders.delete(port);
-    });
-
-    return Promise.resolve(this.#providers[key].meta);
-  }
-
-  public rpcSubscribe ({ method, params, type }: RequestRpcSubscribe, cb: ProviderInterfaceCallback, port: chrome.runtime.Port): Promise<number> {
-    const provider = this.#injectedProviders.get(port);
-
-    assert(provider, 'Cannot call pub(rpc.subscribe) before provider is set');
-
-    return provider.subscribe(type, method, params, cb);
-  }
-
-  public rpcSubscribeConnected (_request: null, cb: ProviderInterfaceCallback, port: chrome.runtime.Port): void {
-    const provider = this.#injectedProviders.get(port);
-
-    assert(provider, 'Cannot call pub(rpc.subscribeConnected) before provider is set');
-
-    cb(null, provider.isConnected()); // Immediately send back current isConnected
-    provider.on('connected', () => cb(null, true));
-    provider.on('disconnected', () => cb(null, false));
-  }
-
-  public rpcUnsubscribe (request: RequestRpcUnsubscribe, port: chrome.runtime.Port): Promise<boolean> {
-    const provider = this.#injectedProviders.get(port);
-
-    assert(provider, 'Cannot call pub(rpc.unsubscribe) before provider is set');
-
-    return provider.unsubscribe(request.type, request.method, request.subscriptionId);
   }
 
   public saveMetadata (meta: MetadataDef): void {
