+++ packages/extension-base/src/background/handlers/types.ts
@@ -1,0 +1,15 @@
+// Copyright 2019-2020 @polkadot/extension authors & contributors
+// This software may be modified and distributed under the terms
+// of the Apache-2.0 license. See the LICENSE file for details.
+
+import { ProviderMeta } from '@polkadot/extension-inject/types';
+import { ProviderInterface } from '@polkadot/rpc-provider/types';
+
+// List of providers passed into constructor. This is the list of providers
+// exposed by the extension.
+export type Providers = Record<string, {
+  meta: ProviderMeta;
+  // The provider is not running at init, calling this will instantiate the
+  // provider.
+  start: () => ProviderInterface;
+}>
